+++ scm-ui/ui-webapp/src/repos/components/list/RepositoryGroupEntry.tsx
@@ -26,17 +26,18 @@ import { Link } from "react-router-dom";
 import { CardColumnGroup, RepositoryEntry } from "@scm-manager/ui-components";
 import { RepositoryGroup } from "@scm-manager/ui-types";
 import { Icon } from "@scm-manager/ui-components";
+import { WithTranslation, withTranslation } from "react-i18next";
 
-type Props = {
+type Props = WithTranslation & {
   group: RepositoryGroup;
 };
 
 class RepositoryGroupEntry extends React.Component<Props> {
   render() {
-    const { group } = this.props;
+    const { group, t } = this.props;
     const settingsLink = group.namespace?._links?.permissions && (
       <Link to={`/namespace/${group.name}/settings`}>
-        <Icon color={"is-link"} name={"cog"} />
+        <Icon color={"is-link"} name={"cog"} title={t("repositoryOverview.settings.tooltip")} />
       </Link>
     );
     const namespaceHeader = (
@@ -54,4 +55,4 @@ class RepositoryGroupEntry extends React.Component<Props> {
   }
 }
 
-export default RepositoryGroupEntry;
+export default withTranslation("namespaces")(RepositoryGroupEntry);