+++ scm-ui/ui-webapp/src/repos/containers/RepositoryRoot.tsx
@@ -54,6 +54,9 @@ import CodeOverview from "../codeSection/containers/CodeOverview";
 import ChangesetView from "./ChangesetView";
 import SourceExtensions from "../sources/containers/SourceExtensions";
 import { FileControlFactory, JumpToFileButton } from "@scm-manager/ui-components";
+import TagsOverview from "../tags/container/TagsOverview";
+import TagRoot from "../tags/container/TagRoot";
+import { urls } from "@scm-manager/ui-components";
 
 type Props = RouteComponentProps &
   WithTranslation & {
@@ -82,25 +85,20 @@ class RepositoryRoot extends React.Component<Props> {
     }
   }
 
-  stripEndingSlash = (url: string) => {
-    if (url.endsWith("/")) {
-      return url.substring(0, url.length - 1);
-    }
-    return url;
-  };
-
-  matchedUrl = () => {
-    return this.stripEndingSlash(this.props.match.url);
-  };
-
   matchesBranches = (route: any) => {
-    const url = this.matchedUrl();
+    const url = urls.matchedUrl(this.props);
     const regex = new RegExp(`${url}/branch/.+/info`);
     return route.location.pathname.match(regex);
   };
 
+  matchesTags = (route: any) => {
+    const url = urls.matchedUrl(this.props);
+    const regex = new RegExp(`${url}/tag/.+/info`);
+    return route.location.pathname.match(regex);
+  };
+
   matchesCode = (route: any) => {
-    const url = this.matchedUrl();
+    const url = urls.matchedUrl(this.props);
     const regex = new RegExp(`${url}(/code)/.*`);
     return route.location.pathname.match(regex);
   };
@@ -118,7 +116,7 @@ class RepositoryRoot extends React.Component<Props> {
 
   evaluateDestinationForCodeLink = () => {
     const { repository } = this.props;
-    const url = `${this.matchedUrl()}/code`;
+    const url = `${urls.matchedUrl(this.props)}/code`;
     if (repository?._links?.sources) {
       return `${url}/sources/`;
     }
@@ -138,7 +136,7 @@ class RepositoryRoot extends React.Component<Props> {
       return <Loading />;
     }
 
-    const url = this.matchedUrl();
+    const url = urls.matchedUrl(this.props);
 
     const extensionProps = {
       repository,
@@ -245,6 +243,15 @@ class RepositoryRoot extends React.Component<Props> {
                   render={() => <BranchesOverview repository={repository} baseUrl={`${url}/branch`} />}
                 />
                 <Route path={`${url}/branches/create`} render={() => <CreateBranch repository={repository} />} />
+                <Route
+                  path={`${url}/tag/:tag`}
+                  render={() => <TagRoot repository={repository} baseUrl={`${url}/tag`} />}
+                />
+                <Route
+                  path={`${url}/tags`}
+                  exact={true}
+                  render={() => <TagsOverview repository={repository} baseUrl={`${url}/tag`} />}
+                />
                 <ExtensionPoint name="repository.route" props={extensionProps} renderAll={true} />
               </Switch>
             </PrimaryContentColumn>
@@ -266,6 +273,16 @@ class RepositoryRoot extends React.Component<Props> {
                   activeWhenMatch={this.matchesBranches}
                   activeOnlyWhenExact={false}
                   title={t("repositoryRoot.menu.branchesNavLink")}
+                />
+                <RepositoryNavLink
+                  repository={repository}
+                  linkName="tags"
+                  to={`${url}/tags/`}
+                  icon="fas fa-tags"
+                  label={t("repositoryRoot.menu.tagsNavLink")}
+                  activeWhenMatch={this.matchesTags}
+                  activeOnlyWhenExact={false}
+                  title={t("repositoryRoot.menu.tagsNavLink")}
                 />
                 <RepositoryNavLink
                   repository={repository}