+++ scm-ui/ui-webapp/src/repos/namespaces/containers/NamespaceRoot.tsx
@@ -33,7 +33,8 @@ import {
   CustomQueryFlexWrappedColumns,
   ErrorPage,
   Loading,
-  Page, PrimaryContentColumn,
+  Page,
+  PrimaryContentColumn,
   SecondaryNavigation,
   SecondaryNavigationColumn,
   StateMenuContextProvider,
@@ -42,6 +43,7 @@ import {
 import Permissions from "../../permissions/containers/Permissions";
 import { ExtensionPoint } from "@scm-manager/ui-extensions";
 import PermissionsNavLink from "./PermissionsNavLink";
+import { urls } from "@scm-manager/ui-components";
 
 type Props = RouteComponentProps &
   WithTranslation & {
@@ -49,6 +51,7 @@ type Props = RouteComponentProps &
     namespaceName: string;
     namespacesLink: string;
     namespace: Namespace;
+    error: Error;
 
     // dispatch functions
     fetchNamespace: (link: string, namespace: string) => void;
@@ -60,20 +63,9 @@ class NamespaceRoot extends React.Component<Props> {
     fetchNamespace(namespacesLink, namespaceName);
   }
 
-  stripEndingSlash = (url: string) => {
-    if (url.endsWith("/")) {
-      return url.substring(0, url.length - 1);
-    }
-    return url;
-  };
-
-  matchedUrl = () => {
-    return this.stripEndingSlash(this.props.match.url);
-  };
-
   render() {
     const { loading, error, namespaceName, namespace, t } = this.props;
-    const url = this.matchedUrl();
+    const url = urls.matchedUrl(this.props);
 
     const extensionProps = {
       namespace,