+++ scm-ui/ui-webapp/src/repos/tags/components/TagButtonGroup.tsx
@@ -22,40 +22,32 @@
  * SOFTWARE.
  */
 
-package sonia.scm.repository;
+import React, { FC } from "react";
+import { Tag, Repository } from "@scm-manager/ui-types";
+import { Button, ButtonAddons } from "@scm-manager/ui-components";
+import { useTranslation } from "react-i18next";
 
-import java.util.Collection;
-import java.util.Optional;
+type Props = {
+  repository: Repository;
+  tag: Tag;
+};
 
-/**
- * Manages namespaces. Mind that namespaces do not have a lifecycle on their own, but only do exist through
- * repositories. Therefore you cannot create or delete namespaces, but just change related settings like permissions
- * associated with them.
- *
- * @since 2.6.0
- */
-public interface NamespaceManager {
+const TagButtonGroup: FC<Props> = ({ repository, tag }) => {
+  const [t] = useTranslation("repos");
 
-  /**
-   * Returns the Namespace with the given name.
-   *
-   * @param namespace The name of the requested namespace.
-   * @return Optional with the namespace for the given name, or an empty Optional if there is no such namespace
-   * (that is, there is no repository with this namespace).
-   */
-  Optional<Namespace> get(String namespace);
+  const changesetLink = `/repo/${repository.namespace}/${repository.name}/code/changeset/${encodeURIComponent(
+    tag.revision
+  )}`;
+  const sourcesLink = `/repo/${repository.namespace}/${repository.name}/sources/${encodeURIComponent(tag.revision)}/`;
 
-  /**
-   * Returns a {@link java.util.Collection} of all namespaces.
-   *
-   * @return all namespaces
-   */
-  Collection<Namespace> getAll();
+  return (
+    <>
+      <ButtonAddons>
+        <Button link={changesetLink} icon="exchange-alt" label={t("tag.commit")} reducedMobile={true} />
+        <Button link={sourcesLink} icon="code" label={t("tag.sources")} reducedMobile={true} />
+      </ButtonAddons>
+    </>
+  );
+};
 
-  /**
-   * Modifies the given namespace.
-   *
-   * @param namespace The namespace to be modified.
-   */
-  void modify(Namespace namespace);
-}
+export default TagButtonGroup;