+++ scm-ui/ui-webapp/src/repos/tags/components/TagView.tsx
@@ -22,40 +22,33 @@
  * SOFTWARE.
  */
 
-package sonia.scm.repository;
+import React, { FC } from "react";
+import { Repository, Tag } from "@scm-manager/ui-types";
+import { ExtensionPoint } from "@scm-manager/ui-extensions";
+import TagDetail from "./TagDetail";
 
-import java.util.Collection;
-import java.util.Optional;
+type Props = {
+  repository: Repository;
+  tag: Tag;
+};
 
-/**
- * Manages namespaces. Mind that namespaces do not have a lifecycle on their own, but only do exist through
- * repositories. Therefore you cannot create or delete namespaces, but just change related settings like permissions
- * associated with them.
- *
- * @since 2.6.0
- */
-public interface NamespaceManager {
+const TagView: FC<Props> = ({ repository, tag }) => {
+  return (
+    <>
+      <TagDetail tag={tag} repository={repository} />
+      <hr />
+      <div className="content">
+        <ExtensionPoint
+          name="repos.tag-details.information"
+          renderAll={true}
+          props={{
+            repository,
+            tag
+          }}
+        />
+      </div>
+    </>
+  );
+};
 
-  /**
-   * Returns the Namespace with the given name.
-   *
-   * @param namespace The name of the requested namespace.
-   * @return Optional with the namespace for the given name, or an empty Optional if there is no such namespace
-   * (that is, there is no repository with this namespace).
-   */
-  Optional<Namespace> get(String namespace);
-
-  /**
-   * Returns a {@link java.util.Collection} of all namespaces.
-   *
-   * @return all namespaces
-   */
-  Collection<Namespace> getAll();
-
-  /**
-   * Modifies the given namespace.
-   *
-   * @param namespace The namespace to be modified.
-   */
-  void modify(Namespace namespace);
-}
+export default TagView;