+++ scm-ui/ui-webapp/src/repos/tags/orderTags.test.ts
@@ -22,40 +22,31 @@
  * SOFTWARE.
  */
 
-package sonia.scm.repository;
+import orderTags from "./orderTags";
 
-import java.util.Collection;
-import java.util.Optional;
+const tag1 = {
+  name: "tag1",
+  revision: "revision1",
+  date: new Date(2020, 1, 1),
+  _links: {}
+};
+const tag2 = {
+  name: "tag2",
+  revision: "revision2",
+  date: new Date(2020, 1, 3),
+  _links: {}
+};
+const tag3 = {
+  name: "tag3",
+  revision: "revision3",
+  date: new Date(2020, 1, 2),
+  _links: {}
+};
 
-/**
- * Manages namespaces. Mind that namespaces do not have a lifecycle on their own, but only do exist through
- * repositories. Therefore you cannot create or delete namespaces, but just change related settings like permissions
- * associated with them.
- *
- * @since 2.6.0
- */
-public interface NamespaceManager {
-
-  /**
-   * Returns the Namespace with the given name.
-   *
-   * @param namespace The name of the requested namespace.
-   * @return Optional with the namespace for the given name, or an empty Optional if there is no such namespace
-   * (that is, there is no repository with this namespace).
-   */
-  Optional<Namespace> get(String namespace);
-
-  /**
-   * Returns a {@link java.util.Collection} of all namespaces.
-   *
-   * @return all namespaces
-   */
-  Collection<Namespace> getAll();
-
-  /**
-   * Modifies the given namespace.
-   *
-   * @param namespace The namespace to be modified.
-   */
-  void modify(Namespace namespace);
-}
+describe("order tags", () => {
+  it("should order tags descending by date", () => {
+    const tags = [tag1, tag2, tag3];
+    orderTags(tags);
+    expect(tags).toEqual([tag2, tag3, tag1]);
+  });
+});