+++ scm-ui/ui-webapp/src/repos/tags/orderTags.ts
@@ -22,40 +22,11 @@
  * SOFTWARE.
  */
 
-package sonia.scm.repository;
+// sort tags by date beginning with latest first
+import { Tag } from "@scm-manager/ui-types";
 
-import java.util.Collection;
-import java.util.Optional;
-
-/**
- * Manages namespaces. Mind that namespaces do not have a lifecycle on their own, but only do exist through
- * repositories. Therefore you cannot create or delete namespaces, but just change related settings like permissions
- * associated with them.
- *
- * @since 2.6.0
- */
-public interface NamespaceManager {
-
-  /**
-   * Returns the Namespace with the given name.
-   *
-   * @param namespace The name of the requested namespace.
-   * @return Optional with the namespace for the given name, or an empty Optional if there is no such namespace
-   * (that is, there is no repository with this namespace).
-   */
-  Optional<Namespace> get(String namespace);
-
-  /**
-   * Returns a {@link java.util.Collection} of all namespaces.
-   *
-   * @return all namespaces
-   */
-  Collection<Namespace> getAll();
-
-  /**
-   * Modifies the given namespace.
-   *
-   * @param namespace The namespace to be modified.
-   */
-  void modify(Namespace namespace);
-}
+export default (tags: Tag[]) => {
+  tags.sort((a, b) => {
+    return new Date(b.date) - new Date(a.date);
+  });
+};